server/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
numpy==1.24.3
opencv-python==4.8.1.78
ultralytics==8.0.196
pymongo==4.5.0
python-dotenv==1.0.0
Pillow==10.0.1
